🌍 #189 Add validation over min/max/origin relationship in Gen.Byte()

+++ tests/GalaxyCheck.Tests.V2/GenTests/ByteGenTests/AboutValueProduction.cs
@@ -13,9 +13,9 @@ namespace Tests.V2.GenTests.ByteGenTests
     {
         [Property]
         public NebulaCheck.IGen<Test> ItProducesValuesLikeInt32() =>
-            from min in Gen.Int32().Between(0, byte.MaxValue)
-            from max in Gen.Int32().Between(min, byte.MaxValue)
-            from origin in Gen.Int32().Between(min, max)
+            from min in Gen.Byte().Between(0, byte.MaxValue)
+            from max in Gen.Byte().Between(min, byte.MaxValue)
+            from origin in Gen.Byte().Between(min, max)
             from bias in DomainGen.Bias()
             from seed in DomainGen.Seed()
             from size in DomainGen.Size()
@@ -24,7 +24,7 @@ namespace Tests.V2.GenTests.ByteGenTests
                 List<object> SampleTraversal(GalaxyCheck.IGen gen) =>
                     gen.Cast<object>().SampleOneTraversal(seed, size);
 
-                var gen0 = GalaxyCheck.Gen.Byte().GreaterThanEqual((byte)min).LessThanEqual((byte)max).ShrinkTowards((byte)origin).WithBias(bias);
+                var gen0 = GalaxyCheck.Gen.Byte().GreaterThanEqual(min).LessThanEqual(max).ShrinkTowards(origin).WithBias(bias);
                 var gen1 = GalaxyCheck.Gen.Int32().Between(min, max).ShrinkTowards(origin).WithBias(bias);
 
                 var sample0 = SampleTraversal(gen0);
@@ -35,30 +35,30 @@ namespace Tests.V2.GenTests.ByteGenTests
 
         [Property]
         public NebulaCheck.IGen<Test> ItProducesValuesGreaterThanOrEqualMinimum() =>
-            from min in Gen.Int32().Between(0, byte.MaxValue)
+            from min in Gen.Byte().Between(0, byte.MaxValue)
             from seed in DomainGen.Seed()
             from size in DomainGen.Size()
             select Property.ForThese(() =>
             {
-                var gen = GalaxyCheck.Gen.Byte().GreaterThanEqual((byte)min);
+                var gen = GalaxyCheck.Gen.Byte().GreaterThanEqual(min);
 
                 var sample = gen.SampleOne(seed: seed, size: size);
 
-                sample.Should().BeGreaterOrEqualTo((byte)min);
+                sample.Should().BeGreaterOrEqualTo(min);
             });
 
         [Property]
         public NebulaCheck.IGen<Test> ItProducesValuesLessThanOrEqualMaximum() =>
-            from max in Gen.Int32().Between(0, byte.MaxValue)
+            from max in Gen.Byte().Between(0, byte.MaxValue)
             from seed in DomainGen.Seed()
             from size in DomainGen.Size()
             select Property.ForThese(() =>
             {
-                var gen = GalaxyCheck.Gen.Byte().LessThanEqual((byte)max);
+                var gen = GalaxyCheck.Gen.Byte().LessThanEqual(max);
 
                 var sample = gen.SampleOne(seed: seed, size: size);
 
-                sample.Should().BeLessOrEqualTo((byte)max);
+                sample.Should().BeLessOrEqualTo(max);
             });
     }
 }
